Introducing Service Layer and Bootstrap 4
Service layer with set of interfaces and their default implementation were introduced. Tests for all new methods were developed first (test driven development) with better separation of concerns between Controller and Service classes.

Bootstrap 4 was introduced for rendering and layout of the various data views. The default list of farms is currently provided in
http://localhost:8080/v1/farms
which is linked to farmer and product profiles. All views are preliminary with images from external sources.